backend/python/services/hair_classification_rag/hair_loss_rag_analyzer_v1_1/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pinecone-client==6.0.0
opencv-python==4.12.0.88
pillow==10.1.0
numpy==2.0.0
scikit-learn==1.3.0
sentence-transformers==3.0.1
torch==2.1.0
torchvision==0.16.0
transformers==4.36.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
python-multipart==0.0.6
openai==1.54.3
